+++ src/main/java/com/lothrazar/cyclicmagic/block/cablepump/TileEntityBasePump.java
@@ -1,4 +1,4 @@
-package com.lothrazar.cyclicmagic.block.pump;
+package com.lothrazar.cyclicmagic.block.cablepump;
 
 import java.util.Map;
 import javax.annotation.Nullable;
