+++ src/main/java/com/lothrazar/cyclicmagic/block/cablepump/fluid/BlockFluidPump.java
@@ -21,33 +21,40 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  ******************************************************************************/
-package com.lothrazar.cyclicmagic.block.pump.fluid;
+package com.lothrazar.cyclicmagic.block.cablepump.fluid;
 
-import com.lothrazar.cyclicmagic.block.cable.TileEntityCableBase;
-import com.lothrazar.cyclicmagic.core.gui.GuiBaseContainer;
-import com.lothrazar.cyclicmagic.gui.GuiSliderInteger;
-import net.minecraft.entity.player.InventoryPlayer;
+import com.lothrazar.cyclicmagic.block.cablepump.BlockPump;
+import com.lothrazar.cyclicmagic.core.IHasRecipe;
+import com.lothrazar.cyclicmagic.gui.ForgeGuiHandler;
+import com.lothrazar.cyclicmagic.registry.RecipeRegistry;
+import net.minecraft.block.ITileEntityProvider;
+import net.minecraft.init.Blocks;
+import net.minecraft.item.ItemStack;
+import net.minecraft.item.crafting.IRecipe;
+import net.minecraft.tileentity.TileEntity;
+import net.minecraft.world.World;
 
-public class GuiFluidPump extends GuiBaseContainer {
+public class BlockFluidPump extends BlockPump implements ITileEntityProvider, IHasRecipe {
 
-  public GuiFluidPump(InventoryPlayer inventoryPlayer, TileEntityFluidPump tileEntity) {
-    super(new ContainerFluidPump(inventoryPlayer, tileEntity), tileEntity);
-    this.fieldRedstoneBtn = TileEntityFluidPump.Fields.REDSTONE.ordinal();
+  public BlockFluidPump() {
+    super();
+    super.setGuiId(ForgeGuiHandler.GUI_INDEX_FLUIDPUMP);
+    this.setFluidTransport();
   }
 
   @Override
-  public void initGui() {
-    super.initGui();
-    int id = 1;
-    int width = 164;
-    int h = 20;
-    int x = this.guiLeft + 6;
-    int y = this.guiTop + 28;
-    //not more than the cable can handle
-    GuiSliderInteger sliderDelay = new GuiSliderInteger(tile, id++, x, y, width, h, 1,
-        TileEntityCableBase.TRANSFER_FLUID_PER_TICK,
-        TileEntityFluidPump.Fields.TRANSFER_RATE.ordinal());
-    sliderDelay.setTooltip("pump.rate");
-    this.addButton(sliderDelay);
+  public TileEntity createNewTileEntity(World worldIn, int meta) {
+    return new TileEntityFluidPump();
+  }
+
+  @Override
+  public IRecipe addRecipe() {
+    return RecipeRegistry.addShapedRecipe(new ItemStack(this),
+        "i i",
+        " r ",
+        "ibi",
+        'b', Blocks.HEAVY_WEIGHTED_PRESSURE_PLATE,
+        'i', "nuggetGold",
+        'r', Blocks.DROPPER);
   }
 }
