+++ src/main/java/com/lothrazar/cyclicmagic/gui/ForgeGuiHandler.java
@@ -47,6 +47,15 @@ import com.lothrazar.cyclicmagic.block.builderpattern.TileEntityPatternBuilder;
 import com.lothrazar.cyclicmagic.block.buildershape.ContainerBuilder;
 import com.lothrazar.cyclicmagic.block.buildershape.GuiBuilder;
 import com.lothrazar.cyclicmagic.block.buildershape.TileEntityStructureBuilder;
+import com.lothrazar.cyclicmagic.block.cablepump.energy.ContainerEnergyPump;
+import com.lothrazar.cyclicmagic.block.cablepump.energy.GuiEnergyPump;
+import com.lothrazar.cyclicmagic.block.cablepump.energy.TileEntityEnergyPump;
+import com.lothrazar.cyclicmagic.block.cablepump.fluid.ContainerFluidPump;
+import com.lothrazar.cyclicmagic.block.cablepump.fluid.GuiFluidPump;
+import com.lothrazar.cyclicmagic.block.cablepump.fluid.TileEntityFluidPump;
+import com.lothrazar.cyclicmagic.block.cablepump.item.ContainerItemPump;
+import com.lothrazar.cyclicmagic.block.cablepump.item.GuiItemPump;
+import com.lothrazar.cyclicmagic.block.cablepump.item.TileEntityItemPump;
 import com.lothrazar.cyclicmagic.block.clockredstone.ContainerClock;
 import com.lothrazar.cyclicmagic.block.clockredstone.GuiClock;
 import com.lothrazar.cyclicmagic.block.clockredstone.TileEntityClock;
@@ -107,15 +116,6 @@ import com.lothrazar.cyclicmagic.block.peat.generator.TileEntityPeatGenerator;
 import com.lothrazar.cyclicmagic.block.placer.ContainerPlacer;
 import com.lothrazar.cyclicmagic.block.placer.GuiPlacer;
 import com.lothrazar.cyclicmagic.block.placer.TileEntityPlacer;
-import com.lothrazar.cyclicmagic.block.pump.energy.ContainerEnergyPump;
-import com.lothrazar.cyclicmagic.block.pump.energy.GuiEnergyPump;
-import com.lothrazar.cyclicmagic.block.pump.energy.TileEntityEnergyPump;
-import com.lothrazar.cyclicmagic.block.pump.fluid.ContainerFluidPump;
-import com.lothrazar.cyclicmagic.block.pump.fluid.GuiFluidPump;
-import com.lothrazar.cyclicmagic.block.pump.fluid.TileEntityFluidPump;
-import com.lothrazar.cyclicmagic.block.pump.item.ContainerItemPump;
-import com.lothrazar.cyclicmagic.block.pump.item.GuiItemPump;
-import com.lothrazar.cyclicmagic.block.pump.item.TileEntityItemPump;
 import com.lothrazar.cyclicmagic.block.screen.ContainerScreen;
 import com.lothrazar.cyclicmagic.block.screen.GuiScreenBlock;
 import com.lothrazar.cyclicmagic.block.screen.TileEntityScreen;
