+++ src/main/java/com/lothrazar/cyclicmagic/module/BlockModule.java
@@ -64,6 +64,12 @@ import com.lothrazar.cyclicmagic.block.cable.item.BlockCableItem;
 import com.lothrazar.cyclicmagic.block.cable.item.TileEntityItemCable;
 import com.lothrazar.cyclicmagic.block.cable.multi.BlockCableBundle;
 import com.lothrazar.cyclicmagic.block.cable.multi.TileEntityCableBundle;
+import com.lothrazar.cyclicmagic.block.cablepump.energy.BlockEnergyPump;
+import com.lothrazar.cyclicmagic.block.cablepump.energy.TileEntityEnergyPump;
+import com.lothrazar.cyclicmagic.block.cablepump.fluid.BlockFluidPump;
+import com.lothrazar.cyclicmagic.block.cablepump.fluid.TileEntityFluidPump;
+import com.lothrazar.cyclicmagic.block.cablepump.item.BlockItemPump;
+import com.lothrazar.cyclicmagic.block.cablepump.item.TileEntityItemPump;
 import com.lothrazar.cyclicmagic.block.clockredstone.BlockRedstoneClock;
 import com.lothrazar.cyclicmagic.block.clockredstone.TileEntityClock;
 import com.lothrazar.cyclicmagic.block.collector.BlockVacuum;
@@ -131,12 +137,6 @@ import com.lothrazar.cyclicmagic.block.peat.generator.BlockPeatGenerator;
 import com.lothrazar.cyclicmagic.block.peat.generator.TileEntityPeatGenerator;
 import com.lothrazar.cyclicmagic.block.placer.BlockPlacer;
 import com.lothrazar.cyclicmagic.block.placer.TileEntityPlacer;
-import com.lothrazar.cyclicmagic.block.pump.energy.BlockEnergyPump;
-import com.lothrazar.cyclicmagic.block.pump.energy.TileEntityEnergyPump;
-import com.lothrazar.cyclicmagic.block.pump.fluid.BlockFluidPump;
-import com.lothrazar.cyclicmagic.block.pump.fluid.TileEntityFluidPump;
-import com.lothrazar.cyclicmagic.block.pump.item.BlockItemPump;
-import com.lothrazar.cyclicmagic.block.pump.item.TileEntityItemPump;
 import com.lothrazar.cyclicmagic.block.scaffolding.BlockScaffolding;
 import com.lothrazar.cyclicmagic.block.scaffolding.BlockScaffoldingReplace;
 import com.lothrazar.cyclicmagic.block.scaffolding.BlockScaffoldingResponsive;
